Move \Slim\Interfaces\CollectionInterface beneath \Slim\Interfaces\Http
